Hit a real web server with a TCP client
