+++ .gitignore
@@ -1,0 +1,7 @@
+.idea/
+/client_recv.txt
+/client_send.txt
+/server_recv.txt
+/server_send.txt
+/.env
+/vendor/
