--- server_recv.txt
@@ -1,1 +1,0 @@
-"My name is Taro"
