--- server_send.txt
@@ -1,1 +1,0 @@
-"Taro,nice to meet you!"
